Pseudocode done
Here is the first fun part of the project, Pseudocode (I have a bit of experience with this from CS50, so If you need to spend a while on this, don't worry, it's normal!). But this is the first fun part, as you're essentially compiling the program mathematically and literally in an English readable format, so you may have to spend some time re-jigging this If when you go through this 'program' it doesn't add up. I believe this is also why they say great programmers can use any stack/language, because the logic is the hard bit, looking up syntax and knowing what to look up is, also hard at first, but is typically the easier bit.