Mystery bug
Here, I must have edited the if, elif, and else statements, as my program started printing an elif statement and the else statement. I'm not too sure what happened, but after scratching my head, I changed all nested if statements to elif, as the second one initially was still if. Not sure If this was the fix, but it seems to be working again now.