Testing user loop
Here I created a new file to avoid complexity for a feature. I'm looking to add the replay feature, and needed to better understand how the while loops worked before adding to my main program, that is starting to build up with visual complexity from the comments (Of which I'm keeping due to the blog. But I will create another version adding all of them together, which will be the final result